Litt omstrukturering og omskriving etter justering av kontantstøtte-kontrakt
