+++ pom.xml
@@ -26,6 +26,7 @@
         <prosessering.version>1.20221110194901_e9e0d90</prosessering.version>
         <felles.version>1.20221101085322_cc4d556</felles.version>
         <kontrakter.version>2.0_20221130105458_c1bd5bb</kontrakter.version>
+<!--        <kontrakter.version>2.0-SNAPSHOT</kontrakter.version>-->
         <main-class>no.nav.familie.baks.mottak.LauncherKt</main-class>
         <spring.cloud.version>3.1.5</spring.cloud.version>
         <token-validation-spring.version>2.1.4</token-validation-spring.version>
@@ -139,6 +140,11 @@
             <groupId>no.nav.familie.kontrakter</groupId>
             <artifactId>barnetrygd</artifactId>
             <version>${kontrakter.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>no.nav.familie.kontrakter</groupId>
+            <artifactId>kontantstotte</artifactId>
+            <version>2.0-SNAPSHOT</version>
         </dependency>
 
         <!-- Kafka -->
