+++ src/main/kotlin/no/nav/familie/baks/mottak/søknad/ArkiverDokumentRequestMapper.kt
@@ -1,5 +1,6 @@
 package no.nav.familie.baks.mottak.søknad
 
+import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.BarnetrygdSøknaddokumentasjon
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.DBSøknad
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.DBVedlegg
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.SøknadV7
@@ -8,14 +9,13 @@ import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.Vedlegg
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.VersjonertSøknad
 import no.nav.familie.baks.mottak.søknad.kontantstøtte.domene.DBKontantstotteVedlegg
 import no.nav.familie.baks.mottak.søknad.kontantstøtte.domene.DBKontantstøtteSøknad
-import no.nav.familie.baks.mottak.søknad.kontantstøtte.domene.KontantstøtteSøknad
+import no.nav.familie.baks.mottak.søknad.kontantstøtte.domene.KontantstøtteSøknaddokumentasjon
 import no.nav.familie.kontrakter.ba.søknad.v4.Søknadstype
-import no.nav.familie.kontrakter.ba.søknad.v7.Dokumentasjonsbehov
-import no.nav.familie.kontrakter.ba.søknad.v7.Søknaddokumentasjon
 import no.nav.familie.kontrakter.felles.dokarkiv.Dokumenttype
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.ArkiverDokumentRequest
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.Dokument
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.Filtype
+import no.nav.familie.kontrakter.ks.søknad.v1.KontantstøtteSøknad
 
 object ArkiverDokumentRequestMapper {
 
@@ -27,8 +27,8 @@ object ArkiverDokumentRequestMapper {
         pdfOriginalSpråk: ByteArray
     ): ArkiverDokumentRequest {
         val (søknadstype, dokumentasjon) = when (versjonertSøknad) {
-            is SøknadV7 -> Pair(versjonertSøknad.søknad.søknadstype, versjonertSøknad.søknad.dokumentasjon)
-            is SøknadV8 -> Pair(versjonertSøknad.søknad.søknadstype, versjonertSøknad.søknad.dokumentasjon)
+            is SøknadV7 -> Pair(versjonertSøknad.søknad.søknadstype, versjonertSøknad.søknad.dokumentasjon.map { BarnetrygdSøknaddokumentasjon(it) })
+            is SøknadV8 -> Pair(versjonertSøknad.søknad.søknadstype, versjonertSøknad.søknad.dokumentasjon.map { BarnetrygdSøknaddokumentasjon(it) })
         }
 
         val dokumenttype = when (søknadstype) {
@@ -97,13 +97,13 @@ object ArkiverDokumentRequestMapper {
             fnr = dbKontantstøtteSøknad.fnr,
             forsøkFerdigstill = false,
             hoveddokumentvarianter = listOf(søknadsdokumentPdf, søknadsdokumentJson),
-            vedleggsdokumenter = hentVedleggListeTilArkivering(kontantstøtteSøknad.dokumentasjon, vedleggMap, pdfOriginalSpråk, Dokumenttype.KONTANTSTØTTE_VEDLEGG),
+            vedleggsdokumenter = hentVedleggListeTilArkivering(kontantstøtteSøknad.dokumentasjon.map { KontantstøtteSøknaddokumentasjon(it) }, vedleggMap, pdfOriginalSpråk, Dokumenttype.KONTANTSTØTTE_VEDLEGG),
             eksternReferanseId = dbKontantstøtteSøknad.id.toString()
         )
     }
 
     private fun hentVedleggListeTilArkivering(
-        dokumentasjon: List<Søknaddokumentasjon>,
+        dokumentasjon: List<ISøknaddokumentasjon>,
         vedleggMap: Map<String, Vedlegg>,
         pdfOriginalSpråk: ByteArray,
         dokumenttype: Dokumenttype
@@ -118,7 +118,7 @@ object ArkiverDokumentRequestMapper {
                             dokument = dbFil.data,
                             dokumenttype = dokumenttype,
                             filtype = Filtype.PDFA,
-                            tittel = dokumentasjonsbehovTilTittel(opplastaVedlegg.tittel)
+                            tittel = opplastaVedlegg.tittel
                         )
                     )
                 }
@@ -138,17 +138,12 @@ object ArkiverDokumentRequestMapper {
 
         return vedlegg
     }
-
-    private fun dokumentasjonsbehovTilTittel(dokumentasjonsbehov: Dokumentasjonsbehov): String {
-        return when (dokumentasjonsbehov) {
-            Dokumentasjonsbehov.ADOPSJON_DATO -> "Adopsjonsdato"
-            Dokumentasjonsbehov.AVTALE_DELT_BOSTED -> "Avtale om delt bosted"
-            Dokumentasjonsbehov.VEDTAK_OPPHOLDSTILLATELSE -> "Vedtak om oppholdstillatelse"
-            Dokumentasjonsbehov.BEKREFTELSE_FRA_BARNEVERN -> "Bekreftelse fra barnevern"
-            Dokumentasjonsbehov.BOR_FAST_MED_SØKER -> "Bor fast med søker"
-            Dokumentasjonsbehov.SEPARERT_SKILT_ENKE -> "Dokumentasjon på separasjon, skilsmisse eller dødsfall"
-            Dokumentasjonsbehov.MEKLINGSATTEST -> "Meklingsattest"
-            Dokumentasjonsbehov.ANNEN_DOKUMENTASJON -> "" // Random dokumentasjon skal saksbehandler sette tittel på
-        }
-    }
 }
+
+interface ISøknaddokumentasjon {
+    val opplastedeVedlegg: List<Søknadsvedlegg>
+}
+data class Søknadsvedlegg(
+    val dokumentId: String,
+    val tittel: String
+)
