+++ src/main/kotlin/no/nav/familie/baks/mottak/søknad/SøknadSpråkvelgerService.kt
@@ -7,7 +7,7 @@ import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.SøknadV7
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.SøknadV8
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.VersjonertSøknad
 import no.nav.familie.baks.mottak.søknad.kontantstøtte.KontantstøtteObjectMapperModule
-import no.nav.familie.baks.mottak.søknad.kontantstøtte.domene.KontantstøtteSøknad
+import no.nav.familie.kontrakter.ks.søknad.v1.KontantstøtteSøknad
 import org.springframework.stereotype.Service
 import no.nav.familie.kontrakter.felles.objectMapper as getObjectMapper
 
@@ -46,6 +46,4 @@ class SøknadSpråkvelgerService {
             KontantstøtteObjectMapperModule(språk)
         )
     }
-
-
 }
