+++ src/main/kotlin/no/nav/familie/baks/mottak/søknad/kontantstøtte/KontantstøtteSøknadController.kt
@@ -3,12 +3,11 @@ package no.nav.familie.baks.mottak.søknad
 import io.micrometer.core.instrument.Metrics
 import no.nav.familie.baks.mottak.søknad.barnetrygd.domene.FødselsnummerErNullException
 import no.nav.familie.baks.mottak.søknad.kontantstøtte.KontantstøtteSøknadService
-import no.nav.familie.baks.mottak.søknad.kontantstøtte.domene.KontantstøtteSøknad
-import no.nav.familie.kontrakter.ba.søknad.v7.Dokumentasjonsbehov
-import no.nav.familie.kontrakter.ba.søknad.v7.Søknaddokumentasjon
-import no.nav.familie.kontrakter.ba.søknad.v7.Søknadsvedlegg
 import no.nav.familie.kontrakter.felles.Ressurs
-import no.nav.security.token.support.core.api.ProtectedWithClaims
+import no.nav.familie.kontrakter.ks.søknad.v1.Dokumentasjonsbehov
+import no.nav.familie.kontrakter.ks.søknad.v1.KontantstøtteSøknad
+import no.nav.familie.kontrakter.ks.søknad.v1.Søknaddokumentasjon
+import no.nav.familie.kontrakter.ks.søknad.v1.Søknadsvedlegg
 import no.nav.security.token.support.core.api.Unprotected
 import org.springframework.http.MediaType.APPLICATION_JSON_VALUE
 import org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE
@@ -21,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping(path = ["/api/kontantstotte/"], produces = [APPLICATION_JSON_VALUE])
-//@ProtectedWithClaims(issuer = "tokenx", claimMap = ["acr=Level4"])
+// @ProtectedWithClaims(issuer = "tokenx", claimMap = ["acr=Level4"])
 @Unprotected
 class KontantstøtteSøknadController(
     private val kontantstøtteSøknadService: KontantstøtteSøknadService
